Update hyprland configuration file
This commit updates the hyprland configuration file to use the correct
file manager and menu along with changes to the general, decoration and
misc section. The styles have been adjusted in the general and
decoration section. The splash has been disabled with the default
wallpaper set to 0 in the misc section.