Update waybar configuration file
This commit updates the waybar configuration files in order to customize
the status bar of the desktop. Both the config.jsonc file and style.css
file has been updated to update the bar settings as well as individual
module settings along with the styles of the bar and the modules.