Add SDDM configuration and Sugar Candy theme configuration
This commit adds the configuration files for SDDM and its theme Sugar
Candy. The SDDM configuration file has been edited so that Sugar Candy
is set as the default theme. The Sugar Candy configuration file have
been edited to change the background image along with the blur and dim
properties.